requirements.txt
mlflow>=2.10
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0

openai>=1.30
tiktoken>=0.7

sentence-transformers>=2.6
transformers>=4.40
torch>=2.2
huggingface-hub>=0.23
tokenizers>=0.19
